requirements.txt
aiogram~=2.25.2
Pillow~=10.1.0
psycopg2~=2.9.9